Add frequency domain seasonal components to UnobservedComponents
Previously, we had implemented the main time domain seasonal component, which
is restrictive when the periodicity is very high or the seasonality can be
more parsimoniously represented by trigonometric terms.  Also, we were only
 able to use a single seasonal component.  This commit allows for multiple
 seasonal components in the frequency domain and for mixing of the two types.

We add a test checking the state matrices for a somewhat
complicated model and update the regression test to be against
the results in the test_ucm.R file as generated by KFAS.

Finally, we added an ipython notebook as an example showcasing the new
functionality.

Also
   - Change offset for regression coefficients when cycle is included.
     Cycle is two states regardless of if it is stochastic.  Therefore,
     the offset should always be 2 instead of possibly 1 for a deterministic
     cycle.
   - Update the .R code for the other seasonal test to correctly initialize P1

Signed-off-by: Jordan Yoder <jordan.yoder@gmail.com>