Only compute matrix_rank(exog) once in DiscreteModel.initialize
I was confused about why a script was taking >20% of total running
time inside DiscreteModel.initialize. It turns out that computing the
rank of exog can be pretty slow! It does seem vary a lot depending on
BLAS and threading settings, e.g. I get wildly different results for
OpenBLAS vs. MKL and OMP_NUM_THREADS=1 versus not. So part of the
issue is probably hitting pathological edge cases in BLAS libraries.

But there's definitely no reason to compute the rank *twice* in the
same function.