Update _smoothers_lowess.pyx
The user is supposed to supply a value in the range [0,1] to the argument frac in the Lowess smoother, but this is not enforced! As a result, if the user mistakenly supplies a value such as frac=40 (for 40%) the software silently defaults to frac=1. Instead, the user should be alerted to this mistake.