Updated durbin_watson Docstring and Tests
Tests for the function durbin_watson function in stats/stattools.py already exist in the stats/tests folder, so I removed the TODO asking for this. Also added new tests which test the function when the input array is multi-dimensional. Also cleaned up the docstring of the function a little.