fix(setup): use correct setuptools backend
The default is `build-backend = "setuptools.build_meta:__legacy__"`, which is not ideal. This field really shouldn't have been optional, but it was.

The legacy backend injects the local path - that has to be done by hand with the non-legacy backend, since packages are discouraged from importing in setup.py. If that makes a difference, that will also have to be added here.